Float fix in checkout page products prices

+++ resources/views/publics/checkout.blade.php
@@ -58,8 +58,8 @@
                         <ul>
                             @foreach($cartProducts as $cartProduct)
                             @php
-                            $sum_total += $cartProduct->num_added * (int)$cartProduct->price;
-                            $sum = $cartProduct->num_added * (int)$cartProduct->price;
+                            $sum_total += $cartProduct->num_added * (float)$cartProduct->price;
+                            $sum = $cartProduct->num_added * (float)$cartProduct->price;
                             @endphp
                             <li>
                                 <input name="id[]" value="{{$cartProduct->id}}" type="hidden">
